Move logs functions from services to utils directory Pass to files new directory of logs functions
