+++ src/db/repository/RetweetLogRepository.ts
@@ -1,5 +1,5 @@
 import models from '../models';
-import logger from '../../services/logs/logger';
+import logger from '../../utils/logs/logger';
 
 import RetweetLogInterface from '../../interfaces/typeDefinitions/RetweetLogInterface';
 import Tweet from '../../interfaces/typeDefinitions/Tweet';
